Experiments with the Homepage
Tried to get the homepage to load the .activity_homepage xml on startup.
It was not having it at all. The main thing I did thats staying is that
I added the scrollable view type in the .activity_homepage so you can
see how that works and how to fit images into it so they scroll.